+++ packages/components/psammead-social-embed/src/Canonical/index.jsx
@@ -1,10 +1,11 @@
+/* eslint-disable no-console */
 import React, { memo, useEffect } from 'react';
 import { func, shape, string } from 'prop-types';
 import styled from '@emotion/styled';
 import useScript from './useScript';
-import fixtures from '../fixtures';
 
 const LANDSCAPE_RATIO = '56.25%';
+const PRE_RENDER_MARGIN = '10rem';
 
 /**
  * Apply provider-specific styles.
@@ -14,6 +15,9 @@ const OEmbed = styled.div`
   display: flex;
   justify-content: center;
 `;
+
+const getOnRenderError = providerName =>
+  `onRender callback function not implemented for ${providerName}`;
 
 /**
  * The following object declares a list of supported Canonical providers
@@ -34,12 +38,16 @@ export const providers = {
         window.instgrm.Embeds.process();
       }
     },
+    onLibraryLoad: () => console.error(getOnRenderError('Instagram')),
   },
   twitter: {
     script: 'https://platform.twitter.com/widgets.js',
     styles: `
       .twitter-tweet {
         margin-top: 0 !important;
+        margin-bottom: ${PRE_RENDER_MARGIN} !important;
+      }
+      .twitter-tweet-rendered {
         margin-bottom: 0 !important;
       }
     `,
@@ -48,7 +56,7 @@ export const providers = {
         window.twttr.widgets.load();
       }
     },
-    onSdkLoad: onRender => {
+    onLibraryLoad: onRender => {
       window.twttr.ready(twttr => {
         twttr.events.bind('rendered', onRender);
       });
@@ -70,18 +78,21 @@ export const providers = {
       }
     `,
     enrich: () => {},
+    onLibraryLoad: () => console.error(getOnRenderError('YouTube')),
   },
 };
 
 const CanonicalEmbed = ({ provider, oEmbed, onRender }) => {
-  const isSdkLoaded = useScript(providers[provider].script);
+  const hasLoadedLibrary = useScript(providers[provider].script);
   useEffect(providers[provider].enrich);
 
   useEffect(() => {
-    if (provider === fixtures.twitter.source && isSdkLoaded && onRender) {
-      providers[fixtures.twitter.source].onSdkLoad(onRender);
+    const { onLibraryLoad } = providers[provider];
+
+    if (onRender && hasLoadedLibrary && onLibraryLoad) {
+      onLibraryLoad(onRender);
     }
-  }, [isSdkLoaded]);
+  }, [hasLoadedLibrary]);
 
   return (
     <OEmbed